requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1